+++ mex/extractors/blueant/checks.py
@@ -7,25 +7,25 @@ from mex.extractors.pipeline.checks.main import (
 
 
 @asset_check(asset="extracted_blueant_activities", blocking=True)
-def check_yaml_rules_exist(context) -> AssetCheckResult:
-    """Check if any asset check rules exist for blueant activity."""
-    extractor= "blueant"
+def check_yaml_rules_exist(context: AssetCheckExecutionContext) -> AssetCheckResult:
+    """Check if any asset check rules exist for blueant activity with valid threshold."""
+    extractor = "blueant"
     entity_type = "activity"
     yaml_exists = check_yaml_path(extractor, entity_type)
     if yaml_exists:
         model = get_rule("x_items_more_than", extractor, entity_type)
         value = model["value"]
-        context.log.info(f"CHECK MODEL: {model}, VALUE: {value}")
         passed = value is not None and value > 0
     else:
         passed = True
-    return AssetCheckResult(
-        passed=passed
-    )
+    return AssetCheckResult(passed=passed)
 
 
-
-@asset_check(asset="extracted_blueant_activities",  additional_deps=["check_yaml_rules_exist"], blocking=True)
+@asset_check(
+    asset="extracted_blueant_activities",
+    additional_deps=["check_yaml_rules_exist"],
+    blocking=True,
+)
 def check_x_items_more_than(
     context: AssetCheckExecutionContext, extracted_blueant_activities: int
 ) -> AssetCheckResult:
@@ -34,6 +34,4 @@ def check_x_items_more_than(
     passed = check_x_items_more_passed(
         context, asset_key, "blueant", "activity", extracted_blueant_activities
     )
-    return AssetCheckResult(
-        passed=passed
-    )
+    return AssetCheckResult(passed=passed)