+++ mex/extractors/pipeline/checks/main.py
@@ -11,10 +11,16 @@ from mex.extractors.pipeline.checks.models.check import AssetCheck
 from mex.extractors.settings import Settings
 from mex.extractors.utils import load_yaml
 
+
 def check_yaml_path(extractor: str, entity_type: str) -> bool:
+    """Checks if there are rules given the extractor and entityType.
+
+    Returns a bool.
+    """
     settings = Settings.get()
     path = settings.all_checks_path / extractor / f"{entity_type}.yaml"
     return path.exists()
+
 
 def load_asset_check_from_settings(extractor: str, entity_type: str) -> AssetCheck:
     """Load AssetCheck model from YAML for a given extractor and entity type."""
@@ -24,7 +30,7 @@ def load_asset_check_from_settings(extractor: str, entity_type: str) -> AssetChe
 
 
 def get_rule(rule: str, extractor: str, entity_type: str) -> dict[str, Any]:
-    """Load rule details from YAML file for given rule type."""
+    """Load rule model from YAML file for given rule type."""
     check_model = load_asset_check_from_settings(
         extractor=extractor, entity_type=entity_type
     )
@@ -32,7 +38,7 @@ def get_rule(rule: str, extractor: str, entity_type: str) -> dict[str, Any]:
 
 
 def parse_time_frame(time_frame: str) -> timedelta:
-    """Parse time frame string like '7d', '3m', '1y' into timedelta."""
+    """Parse time frame string into timedelta."""
     num = int(time_frame[:-1])
     unit = time_frame[-1]
     if unit == "m":
@@ -90,8 +96,10 @@ def check_x_items_more_passed(
     entity_type: str,
     asset_data: int,
 ) -> bool:
-    """Checks rule threshold and returns a bool."""
-    context.log.info(f"CONTEXT: {context}")
+    """Checks whether latest extracted items nr is exceeding the rule threshold.
+
+    Returns bool to AssetCheck.
+    """
     yaml_exists = check_yaml_path(extractor, entity_type)
     if yaml_exists:
         rule = get_rule("x_items_more_than", extractor, entity_type)
@@ -104,15 +112,13 @@ def check_x_items_more_passed(
                 asset_key=asset_key, event_type=DagsterEventType.ASSET_MATERIALIZATION
             )
         )
-        context.log.info(f"EVENTS: {events}")
         historical_events = get_historical_events(events)
         historic_count = get_historic_count(historical_events, time_frame)
         latest_count = asset_data
-        passed =latest_count <= (historic_count if historic_count > 0 else latest_count) + (
-        rule["value"] or 0
-    )
+        passed = latest_count <= (
+            historic_count if historic_count > 0 else latest_count
+        ) + (rule["value"] or 0)
     else:
-        passed= True
+        passed = True
 
-    #context.log.info(f"CHECK MODEL: {rule}, HISTORICAL EVENTS: {historic_count}, LATEST COUNT: {latest_count}")
     return passed