+++ tests/blueant/test_checks.py
@@ -1,4 +1,4 @@
-from datetime import UTC, datetime
+from datetime import UTC, datetime, tzinfo
 from pathlib import Path
 from types import SimpleNamespace
 
@@ -17,7 +17,7 @@ def test_check_yaml_rules_exist_with_real_yaml(monkeypatch: MonkeyPatch) -> None
         all_checks_path = yaml_path
 
         @classmethod
-        def get(cls):
+        def get(cls) -> "MockSettings":
             return cls()
 
     monkeypatch.setattr("mex.extractors.pipeline.checks.main.Settings", MockSettings)
@@ -46,7 +46,7 @@ def test_check_yaml_rules_exist_valid_threshold(
         all_checks_path = yaml_path
 
         @classmethod
-        def get(cls):
+        def get(cls) -> "MockSettings":
             return cls()
 
     monkeypatch.setattr("mex.extractors.pipeline.checks.main.Settings", MockSettings)
@@ -127,7 +127,7 @@ def test_check_yaml_rules_exist_valid_threshold(
         "yaml_not_found_should_pass",
     ],
 )
-def test_check_x_items_more_passed_parametrized(
+def test_check_x_items_more_passed_parametrized(  # noqa: PLR0913
     monkeypatch: MonkeyPatch,
     yaml_exists: False,
     rule_threshold: int,
@@ -140,8 +140,8 @@ def test_check_x_items_more_passed_parametrized(
 
     class FixedDatetime(datetime):
         @classmethod
-        def now(cls, tz=UTC):
-            return mocked_now
+        def now(cls, tz: tzinfo = UTC) -> datetime:
+            return mocked_now.astimezone(tz)
 
     monkeypatch.setattr("mex.extractors.pipeline.checks.main.datetime", FixedDatetime)
 
@@ -157,11 +157,11 @@ def test_check_x_items_more_passed_parametrized(
         )
 
     class DummyMaterialization:
-        def __init__(self, metadata) -> None:
+        def __init__(self, metadata: dict) -> None:
             self.metadata = {"num_items": SimpleNamespace(value=metadata["num_items"])}
 
     class DummyEvent:
-        def __init__(self, timestamp, metadata):
+        def __init__(self, timestamp: datetime, metadata: dict) -> None:
             self.timestamp = timestamp
             self.asset_materialization = DummyMaterialization(metadata)